test: Move byName queries to testing package
Put all query functions into a single module
